+++ project-base/src/Shopsys/ShopBundle/Model/Feed/Exception/FeedConfigNotFoundException.php
@@ -8,11 +8,13 @@ use Shopsys\ShopBundle\Model\Feed\Exception\FeedException;
 class FeedConfigNotFoundException extends Exception implements FeedException
 {
     /**
-     * @param string $message
+     * @param string $feedName
      * @param \Exception|null $previous
      */
-    public function __construct($message = '', Exception $previous = null)
+    public function __construct($feedName = '', Exception $previous = null)
     {
+        $message = 'Feed config with name "' . $feedName . ' not found.';
+
         parent::__construct($message, 0, $previous);
     }
 }